requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
keras==3.2.1
tensorflow==2.16.1
numpy==1.26.4
pillow==10.2.0
python-multipart==0.0.9
